+++ cmd/neofs-node/storage.go
@@ -17,6 +17,7 @@ import (
 	"github.com/nspcc-dev/neofs-node/pkg/util"
 	cid "github.com/nspcc-dev/neofs-sdk-go/container/id"
 	objectSDK "github.com/nspcc-dev/neofs-sdk-go/object"
+	"github.com/nspcc-dev/neofs-sdk-go/user"
 	"github.com/panjf2000/ants/v2"
 	"go.etcd.io/bbolt"
 	"go.uber.org/zap"
@@ -37,12 +38,26 @@ func initLocalStorage(c *cfg) {
 		ls.HandleNewEpoch(ev.(netmap.NewEpoch).EpochNumber())
 	})
 
-	subscribeToContainerRemoval(c, func(id cid.ID) {
-		err := ls.InhumeContainer(id)
-		if err != nil {
-			c.log.Warn("inhuming container after a chain event",
-				zap.Stringer("cID", id), zap.Error(err))
+	subscribeToContainerRemoval(c, func(id cid.ID, owner user.ID) {
+		if owner.IsZero() {
+			err := ls.InhumeContainer(id)
+			if err != nil {
+				c.log.Warn("inhuming container after a chain event",
+					zap.Stringer("cID", id), zap.Error(err))
+			}
+			return
 		}
+		c.log.Info("caught container removal, marking its local objects for GC...",
+			zap.Stringer("container", id), zap.Stringer("owner", owner))
+
+		if err := ls.InhumeContainer(id); err != nil {
+			c.log.Warn("failed to mark local objects from the removed container for GC",
+				zap.Stringer("container", id), zap.Error(err))
+			return
+		}
+
+		c.log.Info("successfully marked local objects from the removed container for GC",
+			zap.Stringer("container", id))
 	})
 
 	// allocate memory for the service;